references/aigc-detector/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
dashscope
httpx
python-dotenv
